requirements.txt
langchain
openai==0.27.8
python-dotenv==1.0.0
streamlit==1.24.1
llama-cpp-python==0.1.65
PyPDF2==3.0.1
tiktoken
qdrant-client==1.3.1
text-generation
sentence_transformers
chromadb